Copy pass for 1.6.0 docs #1.
So: a few changes I've made, which are up for discussion:

* house style seems to be passive voice, so I've removed 'you' where
appropriate
* house style also calls keys on the grid 'keys' or, wherever possible,
just refers to them by the function they are performing ("press
'octave'; the octave key will light"). So I've removed 'button' and made
appropriate changes
* for entirely personal reasons I found 'extended parameters' more
appropriate than 'parameter extensions'
* I've moved some things around. More specifically: the "parameter
extensions" documentation mentions loop/probability/division _before_
we've explained it in the main documentation! So: key, 'basic'
parameters are explained; we mention that there are also extended
parameters, see later; then modifiers are explained; then the extended
parameters section
* replaced h5s with h4s, h5 renders tiny in the default stylesheet
* sub-page, not sub page
* Took a pass at explaining each mode in a single sentence. Also removed
some of the narrative style (referencing previous paras) because
a use-case for documentation is jumping straight to a para. So there's
a bit more repetition, but I think that's OK for technical writing.
* added TODOs for the Teletype commands