small touch-ups to study 0
omg this is doooope. the invocation is really lovely and a perfect summation of the mindset. and the script is super inspiring -- tons to pull out of here immediately into someone's practice.

mostly grammar and spelling, but i also swapped two paragraphs: it seemed useful to turn feedback up before messing with softcut rate modulation so there's stuff in the buffer as learners copy/paste and execute