syntax-y stuff
@fardles , this looks and sounds amazing!! really nice work, thank you. super clear, super playable.

check out `util.wrap`, it's super rad -- it basically does the "add one to our step counter and check if it's over the limit and if so bring the step counter back to the beginning" all in one line.

i also made your `l` variable local, since it doesn't really have meaning outside of the screen redraw function.

other than that, i'm just a jag and made all your 4-tab spacing 2-tab to keep consistent with the rest of the docs :)